+++ crates/networks/src/actor/ext.rs
@@ -11,11 +11,11 @@ pub trait NetworksActorExt {
     // Read operations
     async fn get_current(&self) -> Result<Network>;
     async fn get_list(&self) -> Result<Vec<Network>>;
-    async fn get(&self, chain_id: u32) -> Result<Option<Network>>;
+    async fn get(&self, chain_id: u64) -> Result<Option<Network>>;
     async fn get_by_name(&self, name: String) -> Result<Option<Network>>;
     async fn get_by_id(&self, id: NetworkId) -> Result<Option<Network>>;
-    async fn validate_chain_id(&self, chain_id: u32) -> Result<bool>;
-    async fn get_lowest_dedup_id(&self, chain_id: u32) -> Result<u32>;
+    async fn validate_chain_id(&self, chain_id: u64) -> Result<bool>;
+    async fn get_lowest_dedup_id(&self, chain_id: u64) -> Result<u64>;
 
     // Write operations
     async fn set_current_by_name(&self, new_current_network: String) -> Result<()>;
@@ -35,7 +35,7 @@ impl NetworksActorExt for ActorRef<NetworksActor> {
         Ok(self.ask(GetList).await?)
     }
 
-    async fn get(&self, chain_id: u32) -> Result<Option<Network>> {
+    async fn get(&self, chain_id: u64) -> Result<Option<Network>> {
         Ok(self.ask(Get { chain_id }).await?)
     }
 
@@ -47,11 +47,11 @@ impl NetworksActorExt for ActorRef<NetworksActor> {
         Ok(self.ask(GetById { id }).await?)
     }
 
-    async fn validate_chain_id(&self, chain_id: u32) -> Result<bool> {
+    async fn validate_chain_id(&self, chain_id: u64) -> Result<bool> {
         Ok(self.ask(ValidateChainId { chain_id }).await?)
     }
 
-    async fn get_lowest_dedup_id(&self, chain_id: u32) -> Result<u32> {
+    async fn get_lowest_dedup_id(&self, chain_id: u64) -> Result<u64> {
         Ok(self.ask(GetLowestDedupId { chain_id }).await?)
     }
